Optimizes getting contextual code
