+++ src/shared/types/token.ts
@@ -1,0 +1,10 @@
+export type Token = {
+    line: number;
+    start: number;
+    length: number;
+    tokenTypeNumber: number;
+    tokenType: string | undefined;
+    tokenModifiers: number;
+    name: string;
+    code: string | undefined;
+};
